`_setStream` supports “Unknown” resource type.
The `is_resource` function will return `false` for “Unknown” resource.
This patches does not throw an exception when one tries to set an
resource with such a type. Why? Because when applying some computations
after having disconnected a stream for example, the resource is
considered as “Unknown” but the associated node (its “environment”) is
still valid for computations.

Concrete example with `Hoa\Socket`: broadcasting a message after the
disconnection.