Add the `hasTimedOut` method.
To check whether a connection has timed out, we need to call the
`getStreamMetaData` method which returns an array, and then check the
value of the `timed_out` index in this array. The resulting code is:

    $metaData = $connection->getStreamMetaData();

    if (true === $metaData['timed_out']) {
        // …
    }

In some case, we need to check the time out value very often. In order
to have a better readability of the code, we introduce the `hasTimeOut`
shortcut method. The resulting code is then:

    if (true === $connection->hasTimedOut()) {
        // …
    }

Much better 😀.