Stream: Disallow invalid resource in `_setStream`.
In `_setStream`, we can force a new resource for the stream. There is a
condition checking if the resource is valid. It applies the following rules:

  1. If this is not a resource (i.e. the type is not a resource, OR
     the type is a resource but the resource type is `Unknown`), then
     throw an exception,
  2. If the type is a resource, but the resource type is `Unknown`, do
     NOT throw the exception.

Sometimes, PHP creates “Unknown resources”. This was a bug as far as I
known, so `_setStream` must support “Unknown resources”, they are likely
to be valid in fact. However, they are rejected by `is_resource`. This
is why we need the rule 2.

This specific condition (if X except Y) is not trivial to implement
because of the resource API and wording. If we were giving `true` to
`_setStream`, it was not detected as an invalid resource. This is a
bug. We are responsible of this issue.

This patch fixes it. The new test suite valids this new condition.